tg_bot_crypto/requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
config==0.5.1
configreader==0.0.6
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.8
fonttools==4.48.1
frozendict==2.4.0
frozenlist==1.4.1
html5lib==1.1
idna==3.6
joblib==1.3.2
kiwisolver==1.4.5
lxml==5.1.0
magic-filter==1.0.12
matplotlib==3.8.2
multidict==6.0.5
multitasking==0.0.11
numpy==1.26.4
packaging==23.2
pandas==2.2.0
patsy==0.5.6
peewee==3.17.1
pillow==10.2.0
pmdarima==2.0.4
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
psycopg2
six==1.16.0
soupsieve==2.5
statsmodels==0.14.1
threadpoolctl==3.2.0
SQLAlchemy
SQLAlchemy-JSONField
SQLAlchemy-Utils
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
webencodings==0.5.1
yarl==1.9.4
plotly==5.18.0
psycopg2-binary
yfinance
kaleido